# openapi-schema.json
openapi: 3.0.3
info:
  title: ChargeGhar Nepal API
  version: a315307-dirty
  description: ChargeGhar API - Shared Power Bank Network for Nepal. This documentation
    only shows endpoints documented in our Features TOC.
  contact:
    name: ChargeGhar API Support
    email: support@chargegh.com
  license:
    name: Proprietary
paths:
...
  /api/admin/promotions/analytics:
    get:
      operationId: get_promotion_analytics
      description: Get comprehensive promotion analytics (Staff only)
      summary: Promotion Analytics (Admin)
      tags:
      - Admin
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponAnalyticsResponse'
          description: ''
  /api/admin/promotions/coupons/:
    get:
      operationId: list_admin_coupons
      description: Returns paginated list of all coupons with filtering (Staff only)
      summary: List All Coupons (Admin)
      tags:
      - Admin
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CouponDetail'
          description: ''
    post:
      operationId: create_admin_coupon
      description: Creates new coupon (Staff only)
      summary: Create Coupon (Admin)
      tags:
      - Admin
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CouponCreateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/CouponCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CouponCreateRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponCreate'
          description: ''
  /api/admin/promotions/coupons/{id}/:
    get:
      operationId: get_admin_coupon_details
      description: Retrieves specific coupon details (Staff only)
      summary: Get Coupon Details (Admin)
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this Coupon.
        required: true
      tags:
      - Admin
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponDetail'
          description: ''
    put:
      operationId: update_admin_coupon
      description: Updates coupon information (Staff only)
      summary: Update Coupon (Admin)
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this Coupon.
        required: true
      tags:
      - Admin
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CouponUpdateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/CouponUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CouponUpdateRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponUpdate'
          description: ''
    patch:
      operationId: partial_update_admin_coupon
      description: Partially updates coupon information (Staff only)
      summary: Partial Update Coupon (Admin)
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this Coupon.
        required: true
      tags:
      - Admin
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCouponUpdateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCouponUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCouponUpdateRequest'
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponUpdate'
          description: ''
...
  /api/admin/promotions/coupons/bulk-create/:
    post:
      operationId: api_admin_promotions_coupons_bulk_create_create
      description: Admin-only coupon management ViewSet
      tags:
      - api
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CouponDetailRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/CouponDetailRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CouponDetailRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponDetail'
          description: ''
  /api/admin/promotions/coupons/filter:
    post:
      operationId: filter_admin_coupons
      description: Get filtered and paginated list of coupons (Staff only)
      summary: Filter Coupons (Admin)
      tags:
      - Admin
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CouponFilterRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/CouponFilterRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CouponFilterRequest'
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponFilter'
          description: ''
...
  /api/promotions/coupons/active:
    get:
      operationId: get_active_coupons
      description: Returns list of currently active and valid coupons
      summary: Active Coupons
      tags:
      - Promotions
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActiveCouponsResponse'
          description: ''
  /api/promotions/coupons/apply:
    post:
      operationId: apply_coupon_code
      description: Apply coupon code and receive points
      summary: Apply Coupon Code
      tags:
      - Promotions
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CouponApplyRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/CouponApplyRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CouponApplyRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponApplyResponse'
          description: ''
  /api/promotions/coupons/my:
    get:
      operationId: get_my_coupon_history
      description: Returns user's coupon usage history
      summary: My Coupon History
      tags:
      - Promotions
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyCouponsResponse'
          description: ''
  /api/promotions/coupons/validate:
    post:
      operationId: validate_coupon_code
      description: Check if coupon code is valid and can be used
      summary: Validate Coupon Code
      tags:
      - Promotions
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CouponApplyRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/CouponApplyRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CouponApplyRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponValidationResponse'
          description: ''
...
components:
  schemas:
...
    ActiveCouponsResponse:
      type: object
      description: Response serializer for active coupons endpoint
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/CouponPublic'
        timestamp:
          type: string
          format: date-time
      required:
      - data
      - message
      - success
      - timestamp
...
    CouponAnalyticsResponse:
      type: object
      description: Response serializer for coupon analytics endpoint
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/CouponAnalytics'
        timestamp:
          type: string
          format: date-time
      required:
      - data
      - message
      - success
      - timestamp
...
    CouponApplyRequest:
      type: object
      description: Serializer for applying coupons
      properties:
        coupon_code:
          type: string
          minLength: 1
          maxLength: 10
      required:
      - coupon_code
    CouponApplyResponse:
      type: object
      description: Response serializer for coupon apply endpoint
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/CouponApplyData'
        timestamp:
          type: string
          format: date-time
      required:
      - data
      - message
      - success
      - timestamp
    CouponCreate:
      type: object
      description: Serializer for creating coupons (Admin)
      properties:
        code:
          type: string
          maxLength: 10
        name:
          type: string
          maxLength: 100
        points_value:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        max_uses_per_user:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        status:
          enum:
          - active
          - inactive
          - expired
          type: string
          description: |-
            * `active` - Active
            * `inactive` - Inactive
            * `expired` - Expired
          x-spec-enum-id: 4329db699d4de141
      required:
      - code
      - name
      - points_value
      - valid_from
      - valid_until
    CouponCreateRequest:
      type: object
      description: Serializer for creating coupons (Admin)
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 10
        name:
          type: string
          minLength: 1
          maxLength: 100
        points_value:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        max_uses_per_user:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        status:
          enum:
          - active
          - inactive
          - expired
          type: string
          description: |-
            * `active` - Active
            * `inactive` - Inactive
            * `expired` - Expired
          x-spec-enum-id: 4329db699d4de141
      required:
      - code
      - name
      - points_value
      - valid_from
      - valid_until
    CouponDetail:
      type: object
      description: Detailed serializer for coupons - Full data
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        code:
          type: string
          maxLength: 10
        name:
          type: string
          maxLength: 100
        points_value:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        max_uses_per_user:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        status:
          enum:
          - active
          - inactive
          - expired
          type: string
          description: |-
            * `active` - Active
            * `inactive` - Inactive
            * `expired` - Expired
          x-spec-enum-id: 4329db699d4de141
        created_at:
          type: string
          format: date-time
          readOnly: true
        is_currently_valid:
          type: boolean
          readOnly: true
        days_remaining:
          type: integer
          readOnly: true
        total_uses:
          type: integer
          readOnly: true
      required:
      - code
      - created_at
      - days_remaining
      - id
      - is_currently_valid
      - name
      - points_value
      - total_uses
      - valid_from
      - valid_until
    CouponDetailRequest:
      type: object
      description: Detailed serializer for coupons - Full data
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 10
        name:
          type: string
          minLength: 1
          maxLength: 100
        points_value:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        max_uses_per_user:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        status:
          enum:
          - active
          - inactive
          - expired
          type: string
          description: |-
            * `active` - Active
            * `inactive` - Inactive
            * `expired` - Expired
          x-spec-enum-id: 4329db699d4de141
      required:
      - code
      - name
      - points_value
      - valid_from
      - valid_until
    CouponFilter:
      type: object
      description: Serializer for coupon filtering
      properties:
        status:
          enum:
          - active
          - inactive
          - expired
          type: string
          description: |-
            * `active` - Active
            * `inactive` - Inactive
            * `expired` - Expired
          x-spec-enum-id: 4329db699d4de141
        search:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        page:
          type: integer
          minimum: 1
          default: 1
        page_size:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
    CouponFilterRequest:
      type: object
      description: Serializer for coupon filtering
      properties:
        status:
          enum:
          - active
          - inactive
          - expired
          type: string
          description: |-
            * `active` - Active
            * `inactive` - Inactive
            * `expired` - Expired
          x-spec-enum-id: 4329db699d4de141
        search:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        page:
          type: integer
          minimum: 1
          default: 1
        page_size:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
...
    CouponUpdate:
      type: object
      description: Serializer for updating coupons (Admin)
      properties:
        name:
          type: string
          maxLength: 100
        points_value:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        max_uses_per_user:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        status:
          enum:
          - active
          - inactive
          - expired
          type: string
          description: |-
            * `active` - Active
            * `inactive` - Inactive
            * `expired` - Expired
          x-spec-enum-id: 4329db699d4de141
      required:
      - name
      - points_value
      - valid_from
      - valid_until
    CouponUpdateRequest:
      type: object
      description: Serializer for updating coupons (Admin)
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        points_value:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        max_uses_per_user:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        status:
          enum:
          - active
          - inactive
          - expired
          type: string
          description: |-
            * `active` - Active
            * `inactive` - Inactive
            * `expired` - Expired
          x-spec-enum-id: 4329db699d4de141
      required:
      - name
      - points_value
      - valid_from
      - valid_until
...
    PatchedCouponUpdateRequest:
      type: object
      description: Serializer for updating coupons (Admin)
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        points_value:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        max_uses_per_user:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        status:
          enum:
          - active
          - inactive
          - expired
          type: string
          description: |-
            * `active` - Active
            * `inactive` - Inactive
            * `expired` - Expired
          x-spec-enum-id: 4329db699d4de141
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...